feat: removed optionutil changes and moves the mention calculator to ReportActionCompose
