+++ src/CONST.js
@@ -801,6 +801,9 @@ const CONST = {
         MIN_AMOUNT_OF_ITEMS: 3,
         MAX_AMOUNT_OF_ITEMS: 5,
     },
+    MENTION_SUGGESTER: {
+        HERE_TEXT: '@here',
+    },
     COMPOSER_MAX_HEIGHT: 125,
     CHAT_FOOTER_MIN_HEIGHT: 65,
     CHAT_SKELETON_VIEW: {
