+++ src/pages/home/report/ReportActionCompose.js
@@ -53,7 +53,6 @@ import * as ComposerActions from '../../../libs/actions/Composer';
 import * as Welcome from '../../../libs/actions/Welcome';
 import Permissions from '../../../libs/Permissions';
 import * as TaskUtils from '../../../libs/actions/Task';
-import * as OptionsListUtils from '../../../libs/OptionsListUtils';
 
 const propTypes = {
     /** Beta features list */
@@ -443,50 +442,93 @@ class ReportActionCompose extends React.Component {
     }
 
     /**
+     * Build the suggestions for mentions
+     * @param {Object} personalDetails
+     * @param {String} [searchValue]
+     * @param {Number} [maxRecords]
+     * @returns {Object}
+     */
+    getMentionOptions(personalDetails, searchValue = '', maxRecords = 5) {
+        const suggestions = [];
+
+        if (CONST.MENTION_SUGGESTER.HERE_TEXT.includes(searchValue)) {
+            suggestions.push({
+                text: CONST.MENTION_SUGGESTER.HERE_TEXT,
+                alternateText: this.props.translate('mentionSuggestions.hereAlternateText'),
+                icons: [
+                    {
+                        source: Expensicons.Megaphone,
+                        type: 'avatar',
+                    },
+                ],
+            });
+        }
+        const sortedPersonalDetails = _.sortBy(_.values(personalDetails), (detail) => detail.displayName || detail.login);
+        _.each(sortedPersonalDetails, (detail) => {
+            if (searchValue && !`${detail.displayName} ${detail.login}`.trim().includes(searchValue)) {
+                return;
+            };
+
+            suggestions.push({
+                text: detail.displayName,
+                alternateText: detail.login,
+                icons: [
+                    {
+                        name: detail.login,
+                        source: detail.avatar,
+                        type: 'avatar',
+                    },
+                ]    
+            });
+        });
+        return suggestions.slice(0, maxRecords);
+    }
+
+        /**
      * Clean data related to EmojiSuggestions and MentionSuggestions
      */
-    resetSuggestions() {
-        this.setState({
-            ...this.getDefaultSuggestionsValues(),
-        });
-    }
-
-    /**
-     * Calculates and cares about the content of an Emoji Suggester
-     */
-    calculateEmojiSuggestion() {
-        if (!this.state.value) {
-            this.resetSuggestions();
-            return;
+        resetSuggestions() {
+            this.setState({
+                ...this.getDefaultSuggestionsValues(),
+            });
         }
-        if (this.state.shouldBlockEmojiCalc) {
-            this.setState({shouldBlockEmojiCalc: false});
-            return;
+    
+        /**
+         * Calculates and cares about the content of an Emoji Suggester
+         */
+        calculateEmojiSuggestion() {
+            if (!this.state.value) {
+                this.resetSuggestions();
+                return;
+            }
+            if (this.state.shouldBlockEmojiCalc) {
+                this.setState({shouldBlockEmojiCalc: false});
+                return;
+            }
+            const leftString = this.state.value.substring(0, this.state.selection.end);
+            const colonIndex = leftString.lastIndexOf(':');
+            const isCurrentlyShowingEmojiSuggestion = this.isEmojiCode(this.state.value, this.state.selection.end);
+    
+            // the larger composerHeight the less space for EmojiPicker, Pixel 2 has pretty small screen and this value equal 5.3
+            const hasEnoughSpaceForLargeSuggestion = this.props.windowHeight / this.state.composerHeight >= 6.8;
+            const isAutoSuggestionPickerLarge = !this.props.isSmallScreenWidth || (this.props.isSmallScreenWidth && hasEnoughSpaceForLargeSuggestion);
+    
+            const nextState = {
+                suggestedEmojis: [],
+                highlightedEmojiIndex: 0,
+                colonIndex,
+                shouldShowEmojiSuggestionMenu: false,
+                isAutoSuggestionPickerLarge,
+            };
+            const newSuggestedEmojis = EmojiUtils.suggestEmojis(leftString);
+    
+            if (newSuggestedEmojis.length && isCurrentlyShowingEmojiSuggestion) {
+                nextState.suggestedEmojis = newSuggestedEmojis;
+                nextState.shouldShowEmojiSuggestionMenu = !_.isEmpty(newSuggestedEmojis);
+            }
+    
+            this.setState(nextState);
         }
-        const leftString = this.state.value.substring(0, this.state.selection.end);
-        const colonIndex = leftString.lastIndexOf(':');
-        const isCurrentlyShowingEmojiSuggestion = this.isEmojiCode(this.state.value, this.state.selection.end);
-
-        // the larger composerHeight the less space for EmojiPicker, Pixel 2 has pretty small screen and this value equal 5.3
-        const hasEnoughSpaceForLargeSuggestion = this.props.windowHeight / this.state.composerHeight >= 6.8;
-        const isAutoSuggestionPickerLarge = !this.props.isSmallScreenWidth || (this.props.isSmallScreenWidth && hasEnoughSpaceForLargeSuggestion);
-
-        const nextState = {
-            suggestedEmojis: [],
-            highlightedEmojiIndex: 0,
-            colonIndex,
-            shouldShowEmojiSuggestionMenu: false,
-            isAutoSuggestionPickerLarge,
-        };
-        const newSuggestedEmojis = EmojiUtils.suggestEmojis(leftString);
-
-        if (newSuggestedEmojis.length && isCurrentlyShowingEmojiSuggestion) {
-            nextState.suggestedEmojis = newSuggestedEmojis;
-            nextState.shouldShowEmojiSuggestionMenu = !_.isEmpty(newSuggestedEmojis);
-        }
-
-        this.setState(nextState);
-    }
 
     calculateMentionSuggestion() {
         if (this.state.selection.end < 1) {
@@ -525,7 +567,7 @@ class ReportActionCompose extends React.Component {
         const isCursorBeforeTheMention = valueAfterTheCursor.startsWith(lastWord);
 
         if (!isCursorBeforeTheMention && this.isMentionCode(lastWord)) {
-            const suggestions = OptionsListUtils.getMentionOptions(this.props.personalDetails, prefix);
+            const suggestions = this.getMentionOptions(this.props.personalDetails, prefix);
             nextState.suggestedMentions = suggestions;
             nextState.shouldShowMentionSuggestionMenu = !_.isEmpty(suggestions);
         }
@@ -590,7 +632,7 @@ class ReportActionCompose extends React.Component {
     insertSelectedMention(highlightedMentionIndex) {
         const commentBeforeAtSign = this.state.value.slice(0, this.state.atSignIndex);
         const mentionObject = this.state.suggestedMentions[highlightedMentionIndex];
-        const mentionCode = `@${mentionObject.text === '@here' ? 'here' : mentionObject.alternateText}`;
+        const mentionCode = `@${mentionObject.text === `@${CONST.MENTION_SUGGESTER.HERE_TEXT}` ? CONST.MENTION_SUGGESTER.HERE_TEXT : mentionObject.alternateText}`;
         const commentAfterAtSignWithMentionRemoved = this.state.value.slice(this.state.atSignIndex).replace(CONST.REGEX.MENTION_REPLACER, '');
 
         this.updateComment(`${commentBeforeAtSign}${mentionCode} ${commentAfterAtSignWithMentionRemoved}`, true);
